refactor(shared): 유틸리티와 상수 파일을 shared 폴더로 이동 및 index 파일에서 export 관리
- utils 및 constants 파일을 shared 폴더에 통합
- index.js에서 모든 파일을 export하여 가독성, 접근성 개선
